Agent integrated for MS teams tab
