+++ .env.example
@@ -1,0 +1,19 @@
+# 
+TOPCODER_API_BASE_URL="https://api.topcoder.com/v5"
+AUTH0_M2M_TOKEN_URL="https://auth0proxy.topcoder-dev.com/token" # "https://topcoder-dev.auth0.com/oauth/token"
+AUTH0_M2M_AUDIENCE="https://m2m.topcoder-dev.com/"
+AUTH0_CLIENT_ID=""
+
+# Config for LLM Agent for MS Teams tab app
+# Azure AD SSO Configuration (from your App Registration in Entra ID)
+AZURE_AD_AUDIENCE="api://diamondlike-crosstied-yuette.ngrok-free.app/82d17b02-2d34-4594-b243-09c516aad2e8"
+AZURE_AD_TENANT_ID="9c8f4932-d163-42f7-aed6-b0117730a6e6"
+IS_SAME_AZURE_AD_TENANT=false # true to check loggedIn user tenant should be same
+MOCK_AZURE_AD_VALIDATION=false # true to use mock token instead of Azure AD
+
+AWS_ACCESS_KEY_ID=""
+AWS_SECRET_ACCESS_KEY=""
+AWS_BEDROCK_REGION="us-east-1"
+AWS_BEDROCK_MODEL_ID="anthropic.claude-3-5-sonnet-20240620-v1:0";
+
+MONGO_DB_URL="mongodb://localhost:27017/teams-ai-agent"
