+++ package.json
@@ -21,8 +21,6 @@
   },
   "dependencies": {
     "@aws-sdk/client-bedrock-runtime": "^3.893.0",
-    "@azure/identity": "^4.12.0",
-    "@azure/keyvault-secrets": "^4.10.0",
     "@langchain/aws": "^0.1.15",
     "@langchain/community": "^0.3.56",
     "@langchain/core": "^0.3.77",
@@ -41,6 +39,7 @@
     "jsonwebtoken": "^9.0.2",
     "jwks-rsa": "^3.2.0",
     "langchain": "^0.3.34",
+    "mongoose": "^8.19.1",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
     "uuid": "^9.0.1",
